Added HandleTagNameMixin mixin
Note, "." handling at the end of sting is different from previous
version.
Appending aaa to tag bbb was aaa.bbb, but now it will be aaabbb
Removing aaa from tag aaa.bbb waa bbb, but now it will be .bbb

It was working correctly, but does not compliant with official feature.
So I should call it bug, and fix it!

Sorry for any inconvenient